update [style] addTaskPage
